requirements.txt
pandas
numpy
torch
scikit-learn
mlflow
apache-airflow
kafka-python
flask
requests